Update patch set 3
Patch Set 3:

The person we shooed off was Paul Michali who i've talked with a lot about this library and is above with a +1. I was the most vocal on the other side. 

So this is us in agreement saying that doing things this way covers both our use cases better.

Patch-set: 3
Reviewer: Gerrit User 7191 <7191@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0