Bump up the zake requirement to 0.0.24
Zake was released with a new version with bug fixes
and a addition of the emulation of one more of the
kazoo methods (transaction check support).

This change helps make it possible to test the fix
for bug 1343029 without requiring a full zookeeper
setup to test.

The zake *noteable* changes (for those interested):

- Add travis 3.4 builds
- Add more concurrent transaction tests
- Add much more robust transaction support
- Add basic check transaction support + tests
- Create all parent paths and child paths in the same lock context
- Address concurrent client variable modification
- Move znode making function to storage
- Add link to kazoo readthedocs
- Add test + fix for empty custom versions being provided
- Fix custom version in 'stat' response
- Ensure all paths are normalized before storage interaction

Change-Id: I376bddc78242f2fd559eb80856587c2c4d64b875