Update patch set 2
Patch Set 2:

One addition: If we don't bump the minimum requrement of alembic, it means that neutron db migration healing should work with alembic 0.6.3, but I don't think it is not a right way to address the problem. Even though it is found by third party CI failure, it is not a problem of CI setup.

On the other hand, I agree opinions from Kyle and Clarkb that third party CI should run with the latest version since it is done in OpenStack upstream CI and it would avoid unnecessary differences in running environments.

Patch-set: 2