Update patch set 2
Patch Set 2: Code-Review+1

Yes, my point was that I wasn't going to treat this as an emergency and it should go through normal review. >=0.6.2 does not prevent anyone from installing 0.6.4. So as an interim step you can just install 0.6.4. We should however bump this requirement so this change makes perfect sense. +1 to the change.

This is an important thing that seems to be missed whenever we have these quasi emergency requirements changes. Some changes are actual emergencies and without an updated everything is broken. This was not one of these cases and nothing was broken if you upgraded alembic.

Patch-set: 2
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1