Copy pbr/tools/integration.sh to the requirements repo
The requirements repository is where the integration test is run
from, this is a copy of the pbr integration.sh script to the
repository that actually does the work for validating updates to
the OpenStack global requirements. Housing this script outside
of the requirements repository is a little bit odd.

Once the script is relocated to the requirements repository the
gate and check jobs can start referencing the new location.

Change-Id: Ie1d8415bbfb0c2148352dd7fd61f774ff1502003