Update patch set 1
Patch Set 1:

It's already pinned in stable/icehouse.

taskflow>=0.1.3,<0.2

The version that would bring in wrapt would be 0.4+

Even with the above this is still required due to how the cross release requirements checking occurs. I'm not sure of the entire details (and it's confusing to me as well) but apparently this is required...

Anyways, the upstream creator is still afaik working on the correct fix so this isn't high priority.

Patch-set: 1