Update xstatic-jquery-ui for Horizon
The new versions of jquery-ui have a different layout of files
and directories, which make them incompatible with the old versions.
Rather than pinning the version on ancient 1.8.18, I decided to
make Horizon use the newer version with the newer file layout, hence
the need to upgrade to at least 1.10.1 (current Debian testing).

Change-Id: Ibf599f07b801cd3d907c426486b780aab17ceb9a