Update patch set 1
Patch Set 1: Code-Review+1

I spoke with Steve today that this package is not Python3 compatible and there is a pysaml2-3 package with library implemented for Python3. This is not an issue now as we don't run Keystone with Python3, but we might need to think if two pysaml2 packages might be troublesome or now. Anyway, for now on I think we really need it.

Patch-set: 1
Label: Code-Review=+1