Update patch set 2
Patch Set 2: Code-Review-1

Have you been able to successfully release an alpha without an existing non-alpha version? We had issues with that in Oslo and resorted to 0.1.0 as a first release with alphas only for projects that existed already in Icehouse. Unfortunately, I don't remember the details of the issues or know if they were fixed in infra.

Patch-set: 2