Update patch set 1
Patch Set 1:

It should be noted that devstack gate doesn't have python requirements which is why it isn't gated by gate-requirements. Maybe we should consider making the post job fail on projects liek this more gracefully? but it should be ok to remove it from this list. This wasn't the case before the switch to the complete bandersnatc mirror but it should be fine now with the mirror mirroring all of pypi and being used on all jobs.

Patch-set: 1