Update patch set 7
Patch Set 7:

As Radomir asked on IRC: yes, upgrading to version 1.11 would be a problem in Debian, as it's version 1.10.1 that is in Debian (and there is *A LOT* of reverse dependencies for jquery-ui in Debian). It wouldn't be possible to upgrade it in Sid (though probably upgrading to 1.11 in Debian experimental could be done).

Patch-set: 7