Add attrdict to global requirements
TaskFlow (and likely others) currently has code that allows
a object to be both a dictionary (accessed by key) and a object
that has attributes (accessed by attribute). To allow a nice
reduction of code it would be great to allow for usage of this
library which provides a equivalent object and type instead.

Change-Id: I8a864a4873f00ade4a99b7453598030ec20fd29a