updated requirements repo readme
The purpose of the requirements repository, it's history, and
consistent review guidelines are something that's currently lore
passed down through the ages.

We should write this stuff down, and come up with consistent
guidelines for submitters and reviewers to get us all onto a single
page wrt reviewing and acceptance criteria.

Change-Id: I72f0f5771fee637d0b651d76217ce3959a4b92b7