Update patch set 3
Patch Set 3:

Thanks a lot for the helpful writeup!

Looking at some of the projects listed in projects.txt - and some I even proposed - I wonder whether we need to be more clear on which projects to use: Should it be only projects that are part of a Cloud Installation? Or do we include the tools we use for building those projects as well?

Think about keystone-specs, openstack-manuals and openstack-doc-tools: We currently include them but should we really? For openstack-manuals and openstack-doc-tools, I really like the automatic syncing of requirements but both are kind of developer repos.

Or should we have a second list of projects that like to keep updated on a canonical list but can be less strict?

Patch-set: 3
Label: Verified=0