Add yasfb for making RSS feeds in docs
We are adding an RSS feed to the oslo-specs repository using yasfb
in https://review.openstack.org/120194

This is a documentation build-time dependency, and not a runtime
dependency for OpenStack.

The library isn't tested under Python 3, but I don't foresee any issue
with it working and we don't use Python 3 for building documentation
from Sphinx.

Change-Id: I3d5df90d77a5826c2e5ed463dc957a5477bb3478