Update patch set 1
Patch Set 1:

I'm willing to wait to land this patchset and it's the actual fix in novaclient. This was just the earliest possible moment that I could realistically write these patches (since it took a long time for requests 2.4.x to land with the appropriate version of urllib3. I'll maintain these until the timing is better.

Patch-set: 1
Label: Verified=0