Exclude sqlalchemy-migrate 0.9.2
Currently, sqalchemy-migrate 0.9.2 is causing failures in
Cinder and Nova.  The proposed solution, for the time-being is
to exclude the version of sqlalchemy-migrate which is causing
these issues.

In order to be able to do this we need to update the
global-requirements.  This change has already been made in
master.

Change-Id: I15ddd031e25ac7a39b8a8fea61e3da42a6bd3ac7
Related-bug: 1368391