Update patch set 1
Patch Set 1: Workflow-1

Sounds like there is going to be another release on 9/19 after the oslo libs are updated in global-reqs on 9/18 and there is time to get the requirements files updated in the projects, so they can do their final release (on the 19th).  So I'll update this when we have the final final Juno release version.

Patch-set: 1
Reviewer: Gerrit User 6873 <6873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=-1