Update patch set 1
Patch Set 1:

Hi. It does fix the issue that glanceclient needs to to use urllib3 directly, to keep the same behavior as the packaged version (requests in Debian and Ubuntu uses urllib3, not the requests.packages.urllib3). My intention is to later on change glanceclient to use urllib3 directly as well, but I can't do it without urllib3 being added to the global requirements.

Also, Ian Cordasco, I don't think your -1 is welcome in here. Your position as upstream contributor for requests is in direct conflict of interest for this change, and you voiced a strong opinion in the mailing list that you believe it's OK that requests embeds urllib3. If you want to discuss this fact, please use the list, but don't use the global-requirements.txt as a way to fight your battle. Thanks.

Patch-set: 1
Reviewer: Gerrit User 6476 <6476@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0