Add PyMySQL to global requirements
Theres is a single MySQL connector in a global requirements, so there is
a sense to add one more, at least for py3 testing. Let it be PyMySQL.
PyMySQL is a pure-Python MySQL client library. It hosted on PyPi [1],
licenses by the MIT license, python3 compatible, and seems to be
developing [2].

[1] https://pypi.python.org/pypi/PyMySQL
[2] https://github.com/PyMySQL/PyMySQL/commits/master

Change-Id: If3daeb4559ee3fd7fa69ea654f223ecd13b99cc8