Update patch set 1
Patch Set 1:

Hi Thomas,

I understand but this is needed so that we can tell end-users to upgrade their browsers, so that they can have a good experience with horizon. As You know horizon is the starting point for the end-users and if they will not be notified what is wrong, they will think that our code is broken and it will create a bad experience for them.

Patch-set: 1
Reviewer: Gerrit User 11599 <11599@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0