Update patch set 1
Patch Set 1:

Even if it's necessary, we are still in a deep freeze, and it's too late. Feel free to add the feature *after* the first release of Juno, on a point release. Just now is not the good moment. I'll do the packaging of that new library now though, so we'll see how it goes.

Patch-set: 1