Update patch set 1
Patch Set 1: Code-Review+1

As I wrote in the list, such change would only reflect the current reality with Designate, so it has to be merged for Juno. Also, it's not a problem since Jinja2 is already in Debian and Ubuntu. The only concern is RedHat for which I don't know the state of things.

Patch-set: 1
Reviewer: Gerrit User 6476 <6476@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1