Update patch set 1
Patch Set 1:

This requirement is not in any way related to the production released software, so I don't know if we actually need it. We used to list everything we used in the global requirements list because that was how we got it into the mirrors, but now we are using a full mirror.

Patch-set: 1