suds-jurko - Alternative to suds library for py 3.x support
oslo.vmware would like to support 3.x but suds library is not
maintained and won't work under python 3. suds-jurko is the
most active fork of suds. Moving over projects will be a
step-by-step effort starting with oslo.vmware and then followed
by Nova and Cinder, once the CI jobs stabilize (both our CI and
MineSweeper), we should remove suds entirely from requirements.

We discussed this at the Design summit session for python3 and
there was consensus on this approach.

Change-Id: If91a87ede96a82a2a0d181ee0d4fdc102dc0299d