Update patch set 4
Patch Set 4: Code-Review-1

The trove classifiers for PyKMIP say BSD, but the source does include the Apache license text. Do you know for sure which license is correct?

The library appears to have fairly recent changes, which is good.

The version available on PyPI as I write is 0.1.1. Do you have any idea how far the project is from a 1.0 (or stable API with some other version)?

Patch-set: 4