Update patch set 1
Patch Set 1: Code-Review-1

Now that we have a full pypi mirror available to the CI systems, I think we're trying to switch to only managing dependencies for packages for software products. I'm not certain that the docs need to have their dependencies listed here any more, since the build jobs can install whatever they want and I believe (though am not certain) that the manuals aren't built by the distros. If that assumption is wrong, we should add this, but otherwise I think it's safe for you to just use it without listing it here.

Patch-set: 1