Update patch set 1
Patch Set 1:

Doug, requirements has one nice thing: Keeping versions in sync. And that's the benefit of having this in.

We need this change in openstack-manuals - and since we have openstack-manuals in requirements/projects.txt, we should not add anything new to openstack-manuals/test-requirements.txt AFAIU.

So, what would be the best way of handling this?

Patch-set: 1