Update patch set 4
Patch Set 4:

Our goal is symmetric key store and foundry profile support for PyKMIP v1.0.0. We anticipate this being complete before the end of the Kilo cycle.

We have conducted third-party appliance testing with the library and haven't discovered any blocking issues on our end. As you may know, the library is already in use with Barbican in the SecretStore interface, KMIPSecretStore specifically.

We are actively working on the library. I don't anticipate any breaking of backwards compatibility moving forward; everything we're adding will be an expansion of the KMIP v1.1 spec support, adding to the list of operations and operation subfeatures we currently support.

Patch-set: 4