Update patch set 1
Patch Set 1: Workflow-1

(1 comment)

Thanks for the review. We are the authors of the library also.  The current API available in the library is stable and new API might be added in the future for more functionalities.  We need module 0.1.0 for the drivers in Juno release of Ironic.

Patch-set: 1
Label: Workflow=-1