Update patch set 5
Patch Set 5:

Stefano, I agree with you that slides should not be printed. PDF files are requested by the University Professors which I do not prefer since its not easy to distribute fixes and updates to the slides when students have local copy of the slides in PDF format or have the pages printed.

Patch-set: 5