Added apicapi package to global-requirements.txt
This package is needed for neutron project.
Two modules in neutron project are using this package, but this package
was not added to nuetron project's requirements.txt. So added that to
nuetron project's requirements.txt and submitted the change #129563.
But change #129563 is failing in CI as 'apicapi' package is not found in
global-requirements.txt. So adding 'apicapi' to global-requirements.txt
now.

Change-Id: I550f6987cfea75a1f5e523791a406157503cef0d