Add cm-api to global requirements.
Sahara's CDH plugin requires cm-api[1] library to communicate with
Cloudera Manager (CM). Since it's a vendor specific library, cm-api is
not packaged in Ubuntu or Fedora. This library uses Apache2 license
and is actively maintained by Cloudera [2][3]. cm-api library currently
is the only library that Python code could use to communicate with CM,
and has been used for quite a time in Sahara with good quality
demostrated. According to [1], cm-api has not been validated with
Python 3.0 yet.

[1] https://pypi.python.org/pypi/cm-api/
[2] http://cloudera.github.io/cm_api/docs/python-client/
[3] https://github.com/cloudera/cm_api/tree/master/python

Change-Id: I9e0e19c70437fb3e0fe510fd258d4363e27985c2