Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Overall, yes to this patch, but see inline.

Note this patch itself won't actually freeze mccabe, this adds it to global requirements so we can add it to hacking (so we can release a version of hacking with a pinned version of mccabe). So I think the commit message can use a bit of tweaking to reflect this.

Patch-set: 1
Label: Code-Review=-1