Update patch set 4
Patch Set 4:

I think in terms of stability we do a good job of managing that. In the case of KMIP and PyKMIP there is a testing document that details all of the bytes that must be sent across the wire. With that in hand, it makes it difficult to deviate from the spec. I have been very pleased that because of this it has been a quick process to test on different hardware versions.

http://docs.oasis-open.org/kmip/testcases/v1.1/kmip-testcases-v1.1.pdf

Once a function is implemented I do not anticipate many changes. Our plan is to implement the spec and test with the testing guide. Our final tests are to test against actual hardware.

@Doug What kinds of things are you looking for in terms of stability?

Patch-set: 4