Update patch set 1
Patch Set 1: Code-Review-1

Please add the following (per the readme):

For new Requirements

Is the library actively maintained?

We really want some indication that the library is something we can get support on if we or our users find a bug, and that we don't have to take over and fork the library.

Pointers to recent activity upstream and a consistent release model are appreciated.

Is the library good code?

It's expected, before just telling everyone to download arbitrary 3rd party code from the internet, that the submitter has taken a deep dive into the code to get a feel on whether this code seems solid enough to depend on. That includes ensuring the upstream code has some reasonable testing baked in.

Is the library python 3 compatible?

OpenStack will eventually need to support python 3. At this point adding non python 3 compatible libraries should only be done under extreme need. It should be considered a very big exception.

Is the library license compatible?

Preferably Apache2, BSD, MIT licensed. LGPL is ok. GPL or AGPL is verboten. Any other oddball license should be rejected.

Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?

By adding something to OpenStack global-requirements.txt we are basically demanding that Linux Distros package this for the next release of OpenStack. If they already have, great. If not, we should be cautious of adding it. :ref:`finding-distro-status`

Is the function of this library already covered by other libraries in global-requirements.txt?

Everyone has their own pet libraries that they like to use, but we do not need three different request mocking libraries in OpenStack.

If this new requirement is about replacing an existing library with one that's better suited for our needs, then we also need the transition plan to drop the old library in a reasonable amount of time.

Patch-set: 1