Add voluptuous for libraries/applications data validation
The voluptuous library is much nicer to interact with and
more powerful than the json-schema library (which is more
limited to json) for data-validation.

It would be great to have it avaiable for applications and
libraries in openstack to use for there validation of data
or other inputs/outputs.

The current version in ubuntu trusty is 0.8.2, while fedora
seems to have access to 0.8.5 (so thats why the lower bound
is being set at 0.8.2 in this change).

Change-Id: Ie61e25e968c1a562c9a314a5862220e1919db11b