Add voluptuous for libraries/applications data validation
The voluptuous library *can* be nicer to interact with and
is slightly more powerful (and pythonic) than the json-schema
library for data-validation.

It would be great to have it available for applications and
libraries in openstack to use for there validation of data
or other inputs/outputs (if they so choose to use it over
json-schema).

The current version in ubuntu trusty is 0.8.2, while fedora
seems to have access to 0.8.5 (so thats why the lower bound
is being set at 0.8.2 in this change).

Change-Id: Ie61e25e968c1a562c9a314a5862220e1919db11b