Update patch set 1
Patch Set 1:

I'd like to try using it in taskflow for some messages that are passed back and forth (which currently uses json-schema for basic validation/s); perhaps that isn't enough of a reason to add it though. IMHO it's the least of concerns for openstack to have yet a second/another validation library that may not be json-schema. I'd rather have any type of validation using any of the decent libraries out there (voluptuous seems to belong in this group) rather than no validation at all...

Patch-set: 1