Add cassandra-driver>=2.1.1 requirement
Apache 2 licensed; needed for Cassandra support in Ceilometer

> Is the library actively maintained?

This is the official DataStax library and actively maintains it.
https://github.com/datastax/python-driver

> Is the library good code?

Yes.

> Is the library python 3 compatible?

Yes.

> Is the library license compatible?

Yes: Apache2

> Is the library already packaged in the distros we target (Ubuntu latest /
> Fedora latest)?

Packaged by DataStax, and in PyPi.  Not yet in Ubuntu / Fedora.

However, this is only going to be an _optional_ requirement.  If users do not
enable the Cassandra backend for Ceilometer, they will not use it.  As such,
this is a weaker requirement for packaging OpenStack.

Indeed, it would likely be inappropriate to install the Cassandra driver as part
of a default OpenStack distro, as Cassandra is not a primary datastore for any
OpenStack project.

As such, I do not believe we are requiring the distros to package it.  They may
choose to do so for the convenience of their users, but that is their choice.

> Is the function of this library already covered by other libraries in
> global-requirements.txt?

No

Change-Id: I5803f93be601b875fdfb8276c04e7ecebab8d08e