Update patch set 1
Patch Set 1:

So Fedora and Debian both have patches that make this irrelevant. What's the point of this again? Requests makes no effort maintain compatibility with older versions of urllib3 (nor will we start) so any features that we're using in new versions that aren't present in at least 1.7.1 will cause horrible breakage. Is that what we really want?

Patch-set: 1