Update patch set 1
Patch Set 1: Code-Review+1

Ian, I think you are missing the point. Here, we have glanceclient that is using the embedded version of urllib3 in requests. Yes, we have a fallback to urllib3. But that has been done only because it wasn't possible to add it earlier, since we were in deep freeze of the requirements. The goal here, to add urllib3, is so that Glanceclient can use urllib3 directly, and unconditionally. Indeed, running with the embedded version of urllib3 makes it so that in the gate, we're using a different path than what is in the distributions. This may (and has already in other cases that I had to deal with) lead to false positive tests, which would fail within a distribution environment.

So, we have to add urllib3, then fix glanceclient to use it.

Patch-set: 1
Label: Code-Review=+1