Update patch set 2
Patch Set 2:

I have to disagree.  It pretty much breaks the audit history of stable/*, as a consumer of these branches having a link back to the Master review is pretty important.

I think it would have to be a pretty special case to simply skip this.  The workflow should make it pretty simple to make it accurate.  Am i missing something?

If you feel strongly about this, do you want to raise a debate on openstack-dev to give it to a wider audience?  Thanks

Patch-set: 2