Update patch set 1
Patch Set 1: Code-Review-1

The issue is that unittest2 is pinned below the minimum testtools 1.2.0 w/discover --list needs. Testtools didn't disclose that (but the way we install things in openstack-CI means that wouldn't have shown up anyway).

So this is the wrong end of the lever to push to fix this: unblock unittest2 and things should be fine).

Patch-set: 1
Reviewer: Gerrit User 4190 <4190@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1