Update patch set 1
Patch Set 1: Code-Review-1

The issue is that unittest2 is pinned below the minimum testtools 1.2.0 w/discover --list needs. Testtools didn't disclose that (but the way we install things in openstack-CI means that wouldn't have shown up anyway).

So this is the wrong end of the lever to push to fix this: unblock unittest2 and things should be fine).

Patch-set: 1
Label: Code-Review=-1