Update patch set 1
Patch Set 1:

I misread the error, it is looking for oslo.middleware NOT oslo.messaging.  I am not entirely sure how this has been introduced as a requirement, but I have opened https://review.openstack.org/#/c/134726/ .. but I don't think this is the right way to solve it.

We need to find where this has been referenced.

Patch-set: 1