Update patch set 1
Patch Set 1: -Code-Review

@Sean, we were artificially pinning to an older version of unittest2 because we were installing via distro packages. Testtools didn't declare a minimum version and this issue smoked it out.  They now have declared 'unittest2>=0.8.0' in their recent release.

This issue could just as easily hit a consumer of OpenStack, as we don't necessarily test every combination of deps that requirements.txt allows.

If we were not using distro packages but pypi for every python dep, this issue would not have hit us. The speed that testtools resolved this issue is pretty commendable.

Personally I am not sure this pull request is that important anymore as it creates noise in requirements and that version is only broken if you have an old version of unittest2 (which CI will not now ever have).

There is now a separate effort to try and stop using distro python packages to avoid similar issues in the future.

Patch-set: 1
Reviewer: Gerrit User 979 <979@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0