Update patch set 1
Patch Set 1:

@Alan, This is preciesley that, no? This adds a less than boundary on the last version that is currently suitable for Juno?

The reason this is failing is because oslo.* is currently always installed from Master tip. This needs to be solved before stable/juno requirements can be changed at all.

However, that is a linked but separate issue to this specific pull request. I believe Oslo.messaging is targeting 1.5.0 now, so this probably needs to be updated to be < that.

Patch-set: 1