Update patch set 2
Patch Set 2:

I tend to agree with Sean: as long as the commit message includes the SHA of the original commit, the audit trail is clear enough -- Gerrit even conveniently linkifies it.

I *prefer* when the Change-ID matches, but I don't think that's a valid reason for -1, as long as it was cherry-picked using -x. I would only -1 if there is no way to find the original commit from the proposed change.

Patch-set: 2