Exclude testtools 1.4.0
When running testr with testtools 1.4.0 the regex exclude we use for
the slow tests in tempest fails to work, and the jobs get run on each
run. Let's exclude this release to unblock the gate, we can always
remove the version exclude if it's a bug in tempest's regex or
elsewhere.

Conflicts:
	global-requirements.txt

Change-Id: I561e98bf363d2dd8e8b3a20cea7d2dcbfa034e98
(cherry picked from commit b673b16ac572719cd8552c9cd05448409b628b28)