Update patch set 1
Patch Set 1:

Great, but that's not devstack, so I'd think you'd have your manila devstack add in the manila project on stable/juno and then move that to devstack on master (kilo) and update global-requirements on master (for kilo), which I know you have another change for.  I don't have a problem with the change proposed for master, but I don't think this is appropriate for stable/juno since I don't think the manila change is going to be put into devstack's stable/juno branch.  But I added Dean to keep me honest.

Patch-set: 1