Update patch set 2
Patch Set 2:

Thierry, I take your point and in this example my argument is not particularly strong as it is a simple change.

However, I think it introduces a bad precedent and makes stable reviews harder.  If this sort of practice is done, it then requires the commit messages to be searched for this new proposed valid equivalent to a Change-ID.

This is because the commit-id only links to Master (or original), and it is then not easily indexable which the Change-ID provides of listing all the commits on one click-search.

Ie, in this example which targets stable/icehouse I was unable to quickly check if it had been targeted to stable/juno aswell (this example it didn't need to as it hit Juno whilst it was Master).. but the one way searching causes uncertainty and therefore more work to review or introduced risk.

A potential backporter also has a harder job, as they cannot use the Change-ID as a reference to see if the change has already been proposed or merged, potentially introducing wasted or duplicate work.

It doesn't seem to be too much of a challenge to keep the Change-ID as our valid unique identifier correct, so I am really struggling to support this.

Therefore, I am still keeping my -1.. but that is just my opinion and will I not cry if it is merged.

Patch-set: 2