Update patch set 1
Patch Set 1:

Valeriy,

The point is we don't need to make changes to global-requirements unless something is going to use them, and I'm not convinced that devstack is going to add support for manila on stable/juno. If there is a patch up to devstack that adds that on stable/juno and it's getting agreement from the devstack core team that they want it, but are blocked by this, then I'll remove my -1.

Patch-set: 1