Update patch set 1
Patch Set 1:

Oslo is using semver already. The development version you're seeing there, Alan, is because we have unreleased commits in master after the 1.4.1 release.

We will bump the minor number for each library as we release them during Kilo, so we want to pin all of the Oslo libraries to <major.minor+1.0 (for example, not <=1.4.1 but <1.5.0 for oslo.messaging).

As someone else pointed out, though, the devstack branch in use in the stable branches installs oslo libraries from source and not packages, so pinning the version is not enough to solve the issue.

Patch-set: 1
Reviewer: Gerrit User 2472 <2472@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=0