Update patch set 1
Patch Set 1: Code-Review-1

what session was this discussed in?  I must have missed it.

To clarify, is this just to target projects with a requirements-py3.txt ?

I am really not supportive of introducing a new dependency for general stable/juno requirements.txt, even though it is a fork of one already declared.

However, if it is just for requirements-py3.txt I am more comfortable with it as I would be surprised if anyone is shipping or using py3 with stable/juno for prod.

Currently -1'ing until my concern is addressed.

Thanks.

Patch-set: 1
Reviewer: Gerrit User 979 <979@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1