Update patch set 1
Patch Set 1:

Okay, so simply putting oslo higher in the requirements file didn't seem to help and we're still ending up with eventlet 0.13.0 and the conflict. I'm not sure what else we can do other than cap it here.  AFAICS, the 0.5.0 version's requirements are thin and don't require any other oslo libs that might transitively pull in a requirement on a newer eventlet

Patch-set: 1