Cap Oslo and client library versions
We've agreed to start capping the versions of libraries used in the
stable branches to more accurately reflect what we expect to be used in
a stable deployment, and to allow those libraries to release new
versions
without risking breaking stable deployments.

This patch introduces caps for the Oslo and client libraries to their
current major.minor+1 value (assuming semantic versioning). There will
be another patch to update all of the third-party libraries.

Change-Id: I2d5697871d5750f37399b5b60a92297cbf69d6b5