Update patch set 1
Patch Set 1:

It does have a maintainer and we're working with them on some options for improving the code.

Also, if you want PKI crypto operations in OpenStack - which everyone does, you don't have a huge amount of alternative today. We want to work on Anchor with the community while the (far superior) pycryptography library matures.

Patch-set: 1