Revert "Cap Oslo and client library versions"
Capping the requirements in the stable branch to a version lower than the minimum used in master means we can't do rolling updates any more. We should undo the caps and fix problems when we encounter them (either by fixing the libraries or by capping with some overlap to allow updates).

This reverts commit ad967b1cdb23181de818e9996df7c5462f9ebbec.

Change-Id: Idce155bbb50078d343542560e59b1ef1108f6ea4