Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Just to be clear, the glance tema is in the process of working out the best way to maintain the library from a release and bug-fixing perspective. Whether there's a breaking change or not planned for Kilo, this proposal aims to adopt a new policy that will make future changes easier to maintain.

On the light of what we've done in the Oslo team, I thought that it'd make sense to adopt the same policies we have for Oslo libraries in glance_store. This would allow us to move the development in the master branch forward while still being able to do point releases for bug fixes without shipping new and perhaps incomplete features.

Therefore, this change is proposing to cap the stable branch the same way we do for oslo libraries. A new stable/juno branch will be required for glance_store as well so that we can backport fixes and do point releases from there.

@Erno, small comment in-line

Patch-set: 1
Label: Code-Review=-1