Add aioeventlet dependency
aioeventlet implements the asyncio API (PEP 3156) on top of eventlet:
http://aioeventlet.readthedocs.org/

aioeventlet works on Python 2.6-3.5 and is distributed under the Apache
License 2.0.

I wrote this project to be able to convert incrementally code from
implicit asynchronous programming (eventlet) to explicit asynchronous
programming (asyncio). The project is part of a blueprint to support
asyncio in OpenStack.

The first step in OpenStack is the following change for Oslo Messaging,
"Add a new aioeventlet executor":
https://review.openstack.org/#/c/136653/

A previous try was based on the greenio module, but greenio doesn't fit
well in OpenStack. For the whole history of the integration of asyncio
into OpenStack, read:
http://aioeventlet.readthedocs.org/openstack.html#history

Implements: blueprint greenio-executor
Change-Id: I58cc28f0a6cba4e17741b1987212e711b20eca54