Update patch set 3
Patch Set 3: Code-Review-1

Adding this cap is going to introduce the same problem that we had when we capped all of the libraries. We lose the option of doing rolling updates. We could possibly add a !=1.5.0 setting, but I don't think we want a cap.

Patch-set: 3
Reviewer: Gerrit User 2472 <2472@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1