Switch the SQLAlchemy specifier with something that is more consistent
The previous specifier treater the comma in the syntax as both an
AND and a OR. This sort of worked because of an implementation detail
of setuptools. However setuptools 8.0 (and pip 6.0) will likely
change this implementation detail. Instead use a specifier that treats
the comma consistently and which will work both for older versions
of pip/setuptools and newer ones.

Change-Id: I2f13918c169071d4365c689bdb709a04064279fa