Update patch set 2
Patch Set 2: Code-Review-1

My suggestion is to just set the floor at 0.9.7 in master. It's before milestone 1 so plenty of time to react.

SQLA changes quite a bit between boundaries, and we aren't really testing 0.8.x anywhere at the moment. I think it's time to say we require 0.9 and move on.

Patch-set: 2
Label: Code-Review=-1