Update patch set 1
Patch Set 1: Code-Review-1

I don't have a problem adding these requirements, but I would like a little more info in the commit message about why tooz wants to use a different memcache library. I know we talked about shifting all of the projects over to this library, and this would be a first step, but I'd like to have more detail here in the commit message as an audit trail.

Patch-set: 1
Label: Code-Review=-1