Cap oslo.db for setuptools 8
oslo.db is capped because setuptools 8 changed the way it handles
requirements lists (see 4bbac4c9bc5bcba6962921369ce7721d2a544aa4).
We need to cap at a version known to work with both juno and kilo,
and since 1.2.0 was the last good version known to work we created
1.2.1 with only a requirements change to allow the versions for
juno and kilo to overlap.

Change-Id: Idff40b74d078a786e1322267c500ee952dcb6248