Make SQLAlchemy settings compatible with Juno
We need a new version of oslo.db that works with Juno and Kilo at the same
time, and that has settings for SQLAlchemy that work with setuptools 8.
In I2f13918c169071d4365c689bdb709a04064279fa we removed some of the
overlap with the Juno version requirements, which makes new releases
of oslo.db incompatible with Juno. This change resets the SQLAlchemy
requirements to match Juno.

Change-Id: I3c7301f18bd278da7d8b653a6695c0b4027021b1