Update patch set 1
Patch Set 1: Code-Review-1

We should bump the *minimal* version only if it's a blocker for a project. Not just 'nice to have' or 'we test against it anyway'. The point is to communicate to downstream distributions whether they need to bump their versions, or they are ok to stick to old and proved release if they don't see another reason to bump their versions. The fewer updates to requirements the better.

Patch-set: 1
Reviewer: Gerrit User 9656 <9656@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1