Update patch set 1
Patch Set 1:

Ok dokie.

One piece of the shared code:

https://github.com/openstack/cinder/blob/master/cinder/flow_utils.py 

Glance started to copy this (or part of it), and has an active review that is waiting on being able to avoid that copy @ https://review.openstack.org/#/c/85211/

The newer version of taskflow has this built-in (so the one from cinder can be stripped down to be smaller, and glance doesn't need to copy it in the first place).

That's the current one I can think of; probably others that I'm forgetting to...

Patch-set: 1